sidecar.sh can restart

+++ website/sidecar/sidecar.sh
@@ -6,25 +6,38 @@ LOG=/var/log/sidecar.log
 
 mkdir -p $DATADIR
 
-case $1 in
-    start)
-	cd $CODEDIR
-	nohup node sidecar.js > $LOG 2>&1 &
-	echo started
-    ;;
-    stop)
-	pkill node && echo stopped
-    ;;
-    status)
-	pids=$(pgrep node)
-	if [ -z "$pids" ] ; then
-	    echo No instance of node running
-	else
-	    ps $pids
-	fi
-    ;;
-    *)
-	echo unknown subcommand $1
-	exit 1
-	;;
-esac
+function action() {
+    verb=$1; shift
+    case $verb in
+	start)
+	    cd $CODEDIR
+	    nohup node sidecar.js > $LOG 2>&1 &
+	    echo started
+	    ;;
+	stop)
+	    pids=$(pgrep node)
+	    if [ -n "$pids" ] ; then
+		pkill node && echo stopped
+	    else
+		echo not running
+	    fi
+	    ;;
+	status)
+	    pids=$(pgrep node)
+	    if [ -z "$pids" ] ; then
+		echo No instance of node running
+	    else
+		ps $pids
+	    fi
+	    ;;
+	restart)
+	    action stop
+	    action start
+	*)
+	    echo unknown subcommand $1
+	    exit 1
+	    ;;
+    esac
+}
+
+action "$@"
